+++ packages/vue-component-library/src/stories/BannerFeatured.stories.js
@@ -323,10 +323,10 @@ export function WithBlockForm() {
       return {
         theme: computed(() => ''),
         // explicitly provide a computed property
-        eventId: '9383207',
-        blockFormData: BlockFormData.mock0,
+        eventId: computed(() => '9383207'),
+        blockFormData: computed(() => BlockFormData.mock0),
         libcalEndpoint: 'https://proxy.calendar.library.ucla.edu/',
-        registrationType: 'in-person',
+        registrationType: computed(() => 'in-person'),
       }
     },
     data() {